SUD-26: Adapt sample repo to our package "com.swent.suddenbump"
- Modified folder structure to fit new packages
- Modified imports with new packages
- Added more exclusions in .gitignore
