+++ .gitignore
@@ -1,9 +1,40 @@
-/.idea/
-.DS_Store
-.gradle
-/local.properties
-/build
-/captures
-.externalNativeBuild
-.cxx
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+*.jks
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+replay_pid*
+
+# Config and build files / folders
+.idea/
+.vscode/
+**/build
 local.properties
+keystore.properties
+.gradle
+/app/google-services.json
+/app/release
+upload-keystore.jks
+/scraping/firebase_credentials.json
+
+# MacOS
+**/.DS_Store
