+++ app/src/main/java/com/swent/suddenbump/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.swent.suddenbump.ui.theme
 
 import androidx.compose.ui.graphics.Color
