+++ app/src/main/java/com/swent/suddenbump/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.swent.suddenbump.ui.theme
 
 import android.app.Activity
 import android.os.Build
