+++ app/src/test/java/com/swent/suddenbump/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.swent.suddenbump
 
 import org.junit.Assert.*
 import org.junit.Test
